pyproject.toml
[project]
dependencies = [
    "braceexpand>=0.1.7",
    "cloudpickle>=3.1.1",
    "datasets>=4.3.0",
    "editdistance>=0.8.1",
    "einops>=0.8.1",
    "fiddle>=0.3.0",
    "huggingface-hub<1.0",
    "hydra-core>=1.3.2",
    "ipython>=9.6.0",
    "jiwer>=4.0.0",
    "lhotse>=1.31.1",
    "librosa>=0.11.0",
    "lightning>=2.5.5",
    "matplotlib>=3.10.7",
    "nemo-toolkit>=2.5.2",
    "numpy>=2.3.4",
    "omegaconf>=2.3.0",
    "pandas>=2.3.3",
    "peft>=0.17.1",
    "pyannote-core>=6.0.1",
    "pyannote-metrics>=4.0.0",
    "pynput>=1.8.1",
    "sacrebleu>=2.5.1",
    "sounddevice>=0.5.3",
    "soundfile>=0.13.1",
    "texterrors>=1.0.9",
    "tokenizers>=0.19.1",
    "torch>=2.9.0",
    "transformers>=4.44,<5",
    "wandb>=0.22.3",
    "webdataset>=1.0.2",
    "whisper-normalizer>=0.1.12",
]
description = "Add your description here"
name = "simple-asr"
readme = "README.md"
requires-python = ">=3.12"
version = "0.1.0"

[tool.uv.sources]
torch = { index = "pytorch-cu128" }

[[tool.uv.index]]
explicit = true
name     = "pytorch-cu128"
url      = "https://download.pytorch.org/whl/cu128"

[project.scripts]
start = "simple_asr.main:main"
